pyproject.toml
[tool.poetry]
name = "cgap-gene-annotation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
biopython = "^1.79"
python = "^3.6.2"
black = "^21.9b0"
jsonschema = "3.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pylint = "^2.11"
pytest = "^5.2"

[tool.poetry.scripts]
create_annotation = "cgap_gene_annotation.scripts.create_annotation:main"
update_annotation = "cgap_gene_annotation.scripts.update_annotation:main"
parse_file = "cgap_gene_annotation.scripts.parse_file:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
